Implemented Cache Busting
Using HtmlWebpackPlugin to generate a index.php with a reference to the
newly built js file with a unique hash to prevent the cache holding back
an update from the user.

To do this, I converted the original index.php file to a index.pug file,
which uses pug templating. I couldn't stick to the default template of
HtmlWebpackPlugin, as it used the same formatting style as Backbone.js.
Switching to pug also allowed me to take out all the mitzvah templates
from the file and put it in its own file.

Calling webpack builds the css and js files into one js file, gives it
a unique hash, and rebuilds the index.php file to reference this js
file.

Closed #30.